requirements.txt
Flask==2.3.2
gunicorn==21.2.0
requests==2.31.0
psycopg2-binary==2.9.7
flask-restx==1.3.0
werkzeug==3.1.3